+++ src/common/js/constants.js
@@ -7,6 +7,8 @@ Permissions.C = {
     // verbosity levels
     Verbose: {
         NONE: 0,
-        CONFIGURE: 0x01 <<  0
+        CONFIGURE:   0x01 <<  0,
+        ALLOWED:     0x01 <<  1,
+        NOT_ALLOWED: 0x01 <<  2
     }
 };